Proof of Concept Approved
There was a minor addition for code for the using the IMU w/ the accelerometer, magnetometer, and gyroscope instead of all three separately. Welji checked off on our proof of concept using a PID to react to input from the gyro so the next step is to code and test the IMU. (and figure out how to code for the motors, but one step at a time :D).